+*Use SpV_avg to rescale opacity in non-Bous mode
  Add the optional argument SpV_avg to extract_optics_slice and use it  along
with an appropriate value for opacity_scale to convert the units of opacity from
[Z-1] to [H-1] in non-Boussinesq mode without making use of the Boussinesq
reference density.  Also revised how Hmix_min is set in bulkmixedlayer_init to
avoid any dependency on the Boussinesq reference density.  All Boussinesq
answers are bitwise identical, but non-Boussinesq answers will change and become
less dependent on the Boussinesq reference density, and there is a new optional
argument to a publicly visible subroutine.