+Add BT_RHO_LINEARIZED when non-Boussinesq
  Added the new runtime parameter BT_RHO_LINEARIZED to specify the density that
is used to convert total water column thicknesses into mass in non-Boussinesq
mode with linearized options in the barotropic solver or when estimating the
stable barotropic timestep without access to the full baroclinic model state.
The default is set to RHO_0 and answers do not change by default.  This new
parameter is used in non-Boussinesq mode with some options in btcalc and
find_face_areas, when LINEARIZED_BT_CORIOLIS = True or BT_NONLIN_STRESS = False,
and in the unit conversion of the ice strength with dynamic pressure. Also
modified the units of the gtot_est argument to match those of pbce. There is a
new element in barotropic_CS.  There are also 4 new debugging messages.  All
Boussinesq answers are bitwise identical, and by default non-Boussinesq answers
are as well, but there is a new runtime parameter that will appear in some
MOM_parameter_doc files.