+(*)Minimum non-Boussinesq answer date of 20230701
  Set a minimum answer date of 20230701 when the model is known to be in
non-Boussinesq mode, and ignore all ANSWERS_2018 flags in that same mode.  This
change only applies to the code called from the src directories, and not to code
within or called directly from the config_src/drivers directories or from
another component model like SIS2 where the information about whether the model
is in non-Boussinesq mode is not available.  A verticalGrid_type arguemnt was
added to porous_barriers_init to support one of these changes.  This commit will
change answers in some non-Boussinesq configurations, but all answers in
existing test cases appear to be bitwise identical.  There are fewer entries
logged in non-Boussinesq MOM_parameter_doc files.