*Non-Boussinesq reversion of calculate_CVMix_conv
  Work in height units in calculate_CVMix_conv, including using thickness_to_dz
to convert layer thicknesses to vertical distances.  This eliminates one
instance where GV%Z_to_H was being used in that routine, and is closer to
practice in to what was originally being used on the dev/gfdl branch.  The
answers will change in non-Boussinesq mode when the KPP convection
parameterization is in use, but they are bitwise identical in all Boussinesq
test cases.