*Revise non-Boussinesq wave_speed with no EoS
  Revised wave_speed and wave_speeds to work properly in non-Boussinesq mode
(and not depend on the Boussinesq reference density) when no equation of state
is used.  Also refactored the wave_speed module to reduce code duplication. All
answers are bitwise identical in Boussinesq mode or if an equation of state is
being used, but otherwise answers do change in solutions that depend on the
internal wave speed.