*Revise non-Bous thickness_diffuse with no EoS
  Revised thickness_diffuse_full to work properly in non-Boussinesq mode (and
not depend on the Boussinesq reference density) when no equation of state is
used.  Also refactored this same routine to reduce code duplication. Also added
additional debugging checksums to thickness_diffuse. All answers are bitwise
identical in Boussinesq mode or if an equation of state is being used, but
otherwise answers do change in solutions that use an isopycnal height diffusion
parameterization.