Fix openMP directives in vertvisc_coef
  Corrected two openMP directives in vertvisc_coef to account for recent code
changes related to the nonBoussinesq mode and to the addition of the GL90
parameterization.  Also simplified two openMP directives in vertvisc_limit_vel
and minorly refactored an expression for visc_ml to help ensure that the order
of multiplication of some terms will not allow excessively agressive compilers
to change answers.  All answers are bitwise identical in any cases that worked
before by avoiding openMP threading.