*Restore PE_detrain diagnostic to previous values
  Revised mixedlayer_detrain_2 so that a diagnostic of the potential energy
change implications of detrainment are bitwise identical to their existing
values on the dev/gfdl branch of the MOM6 code.  Without this change, this
diagnostic changes at the level of roundoff in some of the TC test cases.  To
make this change, 4 new internal variables had to be added.  The descriptions of
the scaled units of 8 diagnostics in the TKE budget equation were also
corrected.  The solutions themselves might be bitwise identical.